test: drop duplicate account code

+++ config/account.h
@@ -36,7 +36,7 @@ struct Account
   char *name;                 /**< Name of Account */
   const struct ConfigSet *cs; /**< Parent ConfigSet */
   const char **var_names;     /**< Array of the names of local config items */
-  int num_vars;               /**< Number of local config items */
+  size_t num_vars;            /**< Number of local config items */
   struct HashElem **vars;     /**< Array of the HashElems of local config items */
 };
